pyproject.toml
[project]
name = "johns"
version = "1.0.0"
description = "Discord bot that replies to messages containing 'john'"
requires-python = ">=3.8"
dependencies = ["discord.py>=2.3.0", "python-dotenv>=1.0.1"]

[dependency-groups]
dev = ["ruff>=0.15.5", "ty>=0.0.21"]
